Add parenthesis grouping to scanner
I added the ability to group a lambda calculus expression by parenthesis
finally. As with the tokenizer, it's been hours, and I just want to be
done with the code, so there's not logging, testing, and very minimal
error detection, but I'm pretty sure that it works. At least on the test
file, which can be seen in main.