Add "astification" of lambdas
I added the ability to partially convert lambdas into a curried TreeNode
representation. Because it is only a partial conversion (we don't have
any of the other astification logic in place) I added a placeholder type
to the ParenthesisGrouping enum, ParenthesisGrouping::Processed, that
stores a finished TreeNode enum. I have also changed the type of lambda
bodies in the TreeNode enum, to allow interusage with the newly created
ParenthesisGrouping variant. In the future once all of the astification
logic is in place, I think we will be able to remove this placeholder,
but that will be a while. I also removed the StoredVariable struct from
scanning::tree because I have decided to represent the defined variables
using some sort of hash map instead, and either way the construction of
the final tree is at least a week away, probably more. There is still
lots to do in the middle phase of the parser currently benig developed.
Ultimately, it was just there to cause warnings, and wasn't even going
to be used in the future, so I removed it.